In this commit i replace user methods to one file
