--- src/services/banned.user.service.js
@@ -1,15 +1,0 @@
-import RequestService from './request.service';
-
-const BannedService = {
-    async getAllBanned(){
-        try{
-            const response = await RequestService.get('/api/banned-users');
-            return response;
-        }
-        catch(error){
-            throw new Error(error);
-        }
-    }
-};
-
-export default BannedService;
